requirements.txt
#Έγινε autogeneration από το pip freeze
#Αυτή η εντολή δημιουργεί ένα αρχείο requirements.txt με όλες τις βιβλιοθήκες που είναι εγκατεστημένες στο περιβάλλον.

blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.12.1
python-dotenv==1.1.0
Werkzeug==3.1.3
